refactoring main classes
