+++ microservices/product-composite-service/src/main/java/com/yavlash/microservices/composite/product/config/ApplicationConfiguration.java
@@ -5,13 +5,18 @@ import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Contact;
 import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.info.License;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.client.WebClient;
+import reactor.core.scheduler.Scheduler;
+import reactor.core.scheduler.Schedulers;
 
+@Slf4j
 @Configuration
 @ComponentScan("com.yavlash")
 public class ApplicationConfiguration {
@@ -37,6 +42,25 @@ public class ApplicationConfiguration {
     String apiContactUrl;
     @Value("${api.common.contact.email}")
     String apiContactEmail;
+
+    private final Integer threadPoolSize;
+    private final Integer taskQueueSize;
+
+    @Autowired
+    public ApplicationConfiguration(
+            @Value("${app.threadPoolSize:10}")
+                    Integer threadPoolSize,
+            @Value("${app.taskQueueSize:100}")
+                    Integer taskQueueSize) {
+        this.threadPoolSize = threadPoolSize;
+        this.taskQueueSize = taskQueueSize;
+    }
+
+    @Bean
+    public Scheduler publishEventScheduler() {
+        log.info("Creates a messagingScheduler with connectionPoolSize = {}", threadPoolSize);
+        return Schedulers.newBoundedElastic(threadPoolSize, taskQueueSize, "publish-pool");
+    }
 
     @Bean
     public OpenAPI getOpenApiDocumentation() {
