+++ microservices/recommendation-service/src/main/java/com/yavlash/microservices/core/recommendation/RecommendationServiceApplication.java
@@ -1,19 +1,17 @@
 package com.yavlash.microservices.core.recommendation;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
 
+@Slf4j
 @SpringBootApplication
 public class RecommendationServiceApplication {
-    private static final Logger LOG = LoggerFactory.getLogger(RecommendationServiceApplication.class);
-
     public static void main(String[] args) {
         ConfigurableApplicationContext ctx = SpringApplication.run(RecommendationServiceApplication.class, args);
         String mongodDbHost = ctx.getEnvironment().getProperty("spring.data.mongodb.host");
         String mongodDbPort = ctx.getEnvironment().getProperty("spring.data.mongodb.port");
-        LOG.info("Connected to MongoDb: " + mongodDbHost + ":" + mongodDbPort);
+        log.info("Connected to MongoDb: " + mongodDbHost + ":" + mongodDbPort);
     }
 }
