+++ microservices/review-service/src/main/java/com/yavlash/microservices/core/review/ReviewServiceApplication.java
@@ -1,42 +1,16 @@
 package com.yavlash.microservices.core.review;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import reactor.core.scheduler.Scheduler;
-import reactor.core.scheduler.Schedulers;
 
+@Slf4j
 @SpringBootApplication
-@ComponentScan("com.yavlash")
 public class ReviewServiceApplication {
-    private static final Logger LOG = LoggerFactory.getLogger(ReviewServiceApplication.class);
-    private final Integer threadPoolSize;
-    private final Integer taskQueueSize;
-
-    @Autowired
-    public ReviewServiceApplication(
-            @Value("${app.threadPoolSize:10}") Integer threadPoolSize,
-            @Value("${app.taskQueueSize:100}") Integer taskQueueSize
-    ) {
-        this.threadPoolSize = threadPoolSize;
-        this.taskQueueSize = taskQueueSize;
-    }
-
-    @Bean
-    public Scheduler jdbcScheduler() {
-        LOG.info("Creates a jdbcScheduler with thread pool size = {}", threadPoolSize);
-        return Schedulers.newBoundedElastic(threadPoolSize, taskQueueSize, "jdbc-pool");
-    }
-
     public static void main(String[] args) {
         ConfigurableApplicationContext ctx = SpringApplication.run(ReviewServiceApplication.class, args);
         String mysqlUri = ctx.getEnvironment().getProperty("spring.datasource.url");
-        LOG.info("Connected to MySQL: " + mysqlUri);
+        log.info("Connected to MySQL: " + mysqlUri);
     }
 }
